+++ contracts/src/sentinels/SuperPositionSentinel.sol
@@ -15,14 +15,14 @@ contract SuperPositionSentinel is ISentinel, SuperRegistryImplementer {
     /*//////////////////////////////////////////////////////////////
                                  EVENTS
     //////////////////////////////////////////////////////////////*/
-    event SuperPositionMint(uint32 indexed actionId_, address indexed finalTarget_, uint256 amount_);
-    event SuperPositionBurn(uint32 indexed actionId_, address indexed finalTarget_, uint256 amount_);
+    event SuperPositionMint(uint256 indexed actionId_, address indexed finalTarget_, uint256 amount_);
+    event SuperPositionBurn(uint256 indexed actionId_, address indexed finalTarget_, uint256 amount_);
 
     /*//////////////////////////////////////////////////////////////
                                  EXTERNAL METHODS
     //////////////////////////////////////////////////////////////*/
     /// @inheritdoc ISentinel
-    function notify(uint32 actionId_, address finalTarget_, bytes memory entry_) external {
+    function notify(uint256 actionId_, address finalTarget_, bytes memory entry_) external {
         (uint256 amount, bool mint) = abi.decode(entry_, (uint256, bool));
         if (mint) {
             emit SuperPositionMint(actionId_, finalTarget_, amount);
